doc/sphinx/requirements.txt
# ======= System Dependencies =======
# These must be installed separately using your system package manager:
# - Doxygen (required for API documentation generation)
#   Ubuntu/Debian: sudo apt install doxygen
#   macOS: brew install doxygen
#   RHEL/CentOS/Fedora: sudo dnf install doxygen
# - Graphviz (required for generating diagrams)
#   Ubuntu/Debian: sudo apt install graphviz
#   macOS: brew install graphviz
#   RHEL/CentOS/Fedora: sudo dnf install graphviz
# - LaTeX (required for PDF generation)
#   Ubuntu/Debian: sudo apt install texlive-latex-base texlive-fonts-recommended texlive-latex-extra
#   macOS: brew install --cask mactex
#   RHEL/CentOS/Fedora: sudo dnf install texlive-scheme-medium
# - Inkscape (required for high-quality SVG to PDF conversion)
#   Ubuntu/Debian: sudo apt install inkscape
#   macOS: brew install inkscape
#   RHEL/CentOS/Fedora: sudo dnf install inkscape

# ======= Sphinx =======

# Core Sphinx
sphinx>=5.3,<7                   # Extended version range for Python 3.12
docutils>=0.18                   # More flexible version requirement
Pillow                          # Required for imghdr module
setuptools                      # Required for pkg_resources

# Theme
sphinx-book-theme>=0.4.0         # Book theme (primary)
sphinx_rtd_theme>=1.0            # Read the Docs theme (alternative)

# Extensions used in conf.py
breathe>=4.30                    # C++ documentation integration
myst-parser[linkify]>=0.14.0    # Markdown support
sphinx-design                    # UI components
sphinx-copybutton>=0.5.0         # Copy button for code blocks
sphinx-mermaid                   # Mermaid diagrams support
sphinxcontrib-bibtex>=2.5       # Bibliography support
sphinxcontrib-katex             # Math rendering
standard-imghdr

# Visualization
altair>=5.0                     # Data visualization
sphinxext-altair                # Altair integration

# Optional but useful
sphinx-hoverxref>=0.3b1         # Hover tooltips for references

# Documentation Dependencies
sphinxcontrib-matlabdomain>=0.22.0,<0.23.0  # Pin version for Python 3.12 compatibility
sphinxcontrib-serializinghtml
sphinxcontrib-htmlhelp